requirements.txt
# ===================================================================
# Stock Market Forecasting with GAN - Requirements
# Production-ready Python package dependencies
# ===================================================================

# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Data Collection
yfinance>=0.2.28
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# NLP & Transformers
transformers>=4.35.0
sentencepiece>=0.1.99
accelerate>=0.25.0
tokenizers>=0.15.0

# Machine Learning & Statistics
scikit-learn>=1.3.0
statsmodels>=0.14.0
hmmlearn>=0.3.0  # For regime detection (Hidden Markov Models)

# Experiment Tracking & MLOps
mlflow>=2.8.0
optuna>=3.4.0

# API & Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Configuration Management
pyyaml>=6.0
python-dotenv>=1.0.0
hydra-core>=1.3.0

# Logging & Monitoring
loguru>=0.7.2
tensorboard>=2.14.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Code Quality
black>=23.10.0
flake8>=6.1.0
mypy>=1.6.0
isort>=5.12.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.4.0

# Utilities
tqdm>=4.66.0
python-dateutil>=2.8.2
joblib>=1.3.0

# Database (Optional - for production deployment)
psycopg2-binary>=2.9.9  # PostgreSQL
sqlalchemy>=2.0.0

# Deployment
docker>=6.1.0
gunicorn>=21.2.0
